refactor(schema): split oversized schema modules
